Extract metric doc
